include level information

+++ diag_scripts/aux/C3S_511/lib/get_metadata_to_rst.py
@@ -34,8 +34,8 @@ def do_report(report_data, report_title, work_dir, signature="", latex_opts=Fals
     pid = str(os.getpgid(0))
 
     path_out = work_dir + os.sep + "reporting"    # the final pdf will be put here
-    src_dir = path_out + os.sep + "source_" + pid # Sphinx source code directory
-    bld_dir = path_out + os.sep + "build_" + pid  # Sphinx build directory
+    src_dir = path_out + os.sep + "source_" + report_title + "_" + signature+ "_" + pid # Sphinx source code directory
+    bld_dir = path_out + os.sep + "build_" + report_title + "_" + signature+ "_" + pid  # Sphinx build directory
 
     # create output and temporary directories
     if not (os.path.exists(path_out)):
@@ -88,7 +88,7 @@ def do_report(report_data, report_title, work_dir, signature="", latex_opts=Fals
                     else:
                         outfile.write("* " + key + ": " + report_data["listtext"][key] + "\n")
                         
-                outfile.write(".. raw:: latex \n")
+                outfile.write(".. raw:: latex \n\n")
                 outfile.write("   \clearpage \n") #does not react on this
                         
             else:
@@ -122,8 +122,8 @@ def do_report(report_data, report_title, work_dir, signature="", latex_opts=Fals
             else:
                 print("Wrong format in text entry, nothing can be written!") # TODO: ERROR function 
                 
-#            outfile.write(".. raw:: latex \n")
-#            outfile.write("   \clearpage \n") #does not react on this
+            outfile.write(".. raw:: latex \n\n")
+            outfile.write("   \clearpage \n") #does not react on this
                 
         else:
             print("No writable text found! There was no 'freetext' in the dictionary!")
@@ -158,6 +158,8 @@ def do_report(report_data, report_title, work_dir, signature="", latex_opts=Fals
                                   "   :width:   95%" + "\n\n" 
                                   "   " + caption[0] + "\n"
                                  )
+                    outfile.write(".. raw:: latex \n\n")
+                    outfile.write("   \FloatBarrier \n") #does not react on this
         
             else:
                 print("Wrong format in plots entry, nothing can be written!") # TODO: ERROR function 
@@ -165,7 +167,7 @@ def do_report(report_data, report_title, work_dir, signature="", latex_opts=Fals
         else:
             print("No plottable links found! There was no 'plots' in the dictionary!")
             
-        if "text" not in report_data.keys() and "plots" not in report_data.keys():
+        if "freetext" not in report_data.keys() and "listtext" not in report_data.keys() and "plots" not in report_data.keys():
             print("Nothing to write reports from!! There was no 'plots' nor 'text' in the dictionary!") # TODO: ERROR function 
             return
 
@@ -179,34 +181,35 @@ def do_report(report_data, report_title, work_dir, signature="", latex_opts=Fals
     os.environ['SOURCEDIR'] = src_dir
     os.environ['BUILDDIR'] = bld_dir
 
-    # run Sphinx to create a pdf
-    oldpath = os.getcwd()
-    os.chdir("doc/reporting")
     
-    if not latex_opts:
-        with open(os.devnull, 'wb') as devnull:
-            subprocess.call("make latexpdf", shell=True,
-                            stdout=devnull, stderr=subprocess.STDOUT)
-    else:
-        subprocess.call("make latexpdf", shell=True)
+    if latex_opts is not None:
+        # run Sphinx to create a pdf
+        oldpath = os.getcwd()
+        os.chdir("doc/reporting")
+        if not latex_opts:
+            with open(os.devnull, 'wb') as devnull:
+                subprocess.call("make latexpdf", shell=True,
+                                stdout=devnull, stderr=subprocess.STDOUT)
+        else:
+            subprocess.call("make latexpdf", shell=True)
          
-    os.chdir(oldpath)
+        os.chdir(oldpath)
 
-    # move pdf to the output directory and rename to report_xxx.pdf
-    pdfname = path_out + os.sep + "report_" + report_title.split()[0].lower() + "_" + signature + ".pdf"
-    os.rename(bld_dir + os.sep + "latex" + os.sep + "ESMValToolC3S_511Report.pdf", pdfname)
+        # move pdf to the output directory and rename to report_xxx.pdf
+        pdfname = path_out + os.sep + "report_" + report_title.split()[0].lower() + "_" + signature + ".pdf"
+        os.rename(bld_dir + os.sep + "latex" + os.sep + "ESMValToolC3S_511Report.pdf", pdfname)
 
-    # clean up temporary directories
-    if os.path.exists(src_dir):
-         # remove if exists
-         shutil.rmtree(src_dir)
-         pass
-    if os.path.exists(bld_dir):
-         # remove if exists
-         shutil.rmtree(bld_dir)
-         pass
-
-    print("Successfully created " + pdfname + "!")
+        # clean up temporary directories
+        if os.path.exists(src_dir):
+             # remove if exists
+             shutil.rmtree(src_dir)
+             pass
+        if os.path.exists(bld_dir):
+             # remove if exists
+             shutil.rmtree(bld_dir)
+             pass
+    
+        print("Successfully created " + pdfname + "!")
     
     
 def do_smm_table(csv_expert, csv_definitions):
